+++ src/main/java/com/dnd/modutime/controller/GlobalControllerAdvice.java
@@ -1,6 +1,7 @@
 package com.dnd.modutime.controller;
 
-import com.dnd.modutime.application.InvalidPasswordException;
+import com.dnd.modutime.exception.InvalidPasswordException;
+import com.dnd.modutime.exception.NotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -11,8 +12,20 @@ import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExcep
 public class GlobalControllerAdvice extends ResponseEntityExceptionHandler {
 
     @ExceptionHandler(InvalidPasswordException.class)
-    public ResponseEntity<ExceptionResponse> unAuthorizedExceptionHandler(InvalidPasswordException exception) {
+    public ResponseEntity<ExceptionResponse> handleUnAuthorizedException(InvalidPasswordException exception) {
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body(new ExceptionResponse(exception.getMessage()));
+    }
+
+    @ExceptionHandler(NotFoundException.class)
+    public ResponseEntity<ExceptionResponse> handleNotFoundException(NotFoundException exception) {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body(new ExceptionResponse(exception.getMessage()));
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<ExceptionResponse> handleIllegalArgumentException(IllegalArgumentException exception) {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                 .body(new ExceptionResponse(exception.getMessage()));
     }
 }
