+++ src/main/java/com/dnd/modutime/exception/InvalidPasswordException.java
@@ -1,4 +1,4 @@
-package com.dnd.modutime.application;
+package com.dnd.modutime.exception;
 
 public class InvalidPasswordException extends RuntimeException {
     public InvalidPasswordException() {
