+++ src/main/java/com/dnd/modutime/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package com.dnd.modutime.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
